+++ src/rocket-pack/packageJson/computePackageJson.ts
