+++ src/rocket-pack/packageJson/getExternalPackages.ts
