+++ src/rocket-pack/packageJson/getInternalPackages.ts
