+++ src/rocket-pack/rule/readDirectoryPatterns.ts
